// Lab02-Desenv-Software/implementação/frontEnd/car-rental-frontend/src/utils/enums/Enum.tsx
export enum StatusDocumento {
    APROVADO  = "APROVADO",
    REPROVADO = "REPROVADO",
    PENDENTE  = "PENDENTE"
}

export enum DuracaoContrato {
    DOZE_MESES = 12,
    VINTE_QUATRO_MESES = 24,
    TRINTA_SEIS_MESES = 36,
    QUARENTA_OITO_MESES = 48,
}
  
export enum TipoContrato {
    SEGURO  = "SEGURO",
    ALUGUEL = "ALUGUEL",
    COMPRA  = "COMPRA",
    CREDITO = "CREDITO"
}

export enum TipoAgente {
    EMPRESA = "EMPRESA",
    BANCO = "BANCO"
}